Add relic XP view to overlay, refactor relic XP view
