+++ ICE/Ui/SettngsWindowV2.cs
@@ -843,6 +843,7 @@ internal class SettingsWindowV2 : Window
 
     private bool showOverlay = C.ShowOverlay;
     private bool ShowSeconds = C.ShowSeconds;
+    private bool ShowXP = C.ShowXP;
 
     private void Overlay()
     {
@@ -855,6 +856,12 @@ internal class SettingsWindowV2 : Window
         if (ImGui.Checkbox("Show Seconds", ref ShowSeconds))
         {
             C.ShowSeconds = ShowSeconds;
+            C.Save();
+        }
+
+        if (ImGui.Checkbox("Show XP", ref ShowXP))
+        {
+            C.ShowXP = ShowXP;
             C.Save();
         }
     }
